fix: improve first-time user experience
- Fix postinstall typo: `create` → `init` command
- Add `createChatHandler` factory to eliminate ~130 lines of duplicated
  boilerplate per chat route template (Zod schemas, message transform,
  error handling)
- Update 5 template routes to use the new one-liner factory
- Fix misleading config reference in about.mdx (config is optional)
- Make web-search stub throw actionable error instead of returning fake data
- Regenerate template manifest
