+++ cli/templates/files/minimal/app/about/page.mdx
@@ -15,7 +15,7 @@ This is a minimal starter template.
 
 1. Edit pages in the `app` directory
 2. Add components in `components`
-3. Configure your app in `veryfront.config.js`
+3. Customize with `veryfront.config.ts` when needed
 
 Happy coding!
