+++ cli/templates/files/multi-agent/tools/web-search.ts
@@ -7,18 +7,12 @@ export default tool({
   inputSchema: z.object({
     query: z.string().describe("Search query"),
   }),
-  execute: async ({ query }) => {
-    // Replace with a real search API (e.g., Tavily, SerpAPI, Brave Search)
-    return {
-      results: [
-        {
-          title: `Results for: ${query}`,
-          snippet:
-            "Connect a search API to get real results. " +
-            "See https://veryfront.com/code/guides/tools for setup instructions.",
-          url: "https://veryfront.com/code/guides/tools",
-        },
-      ],
-    };
+  execute: async ({ query: _query }) => {
+    // Connect a real search API to use this tool.
+    // Popular options: Tavily, SerpAPI, Brave Search
+    throw new Error(
+      "No search API configured. " +
+        "See https://veryfront.com/code/guides/tools for setup instructions.",
+    );
   },
 });
