+++ src/agent/index.ts
@@ -122,4 +122,5 @@ export {
 } from "./composition/index.ts";
 
 export { agent } from "./factory.ts";
+export { type ChatHandlerOptions, createChatHandler } from "./chat-handler.ts";
 export { AgentRuntime } from "./runtime/index.ts";
